Devreverted

+++ backend/api/controllers/recordsController.go
@@ -13,6 +13,16 @@ import (
 
 func RecordCreate(c *fiber.Ctx) error {
 	var record models.Record
+
+	userCredentials := c.Locals("user").(fiber.Map)
+	role := userCredentials["Role"].(string)
+
+	if role != "tutor" {
+		return c.Status(http.StatusUnauthorized).JSON(fiber.Map{
+			"error": "unauthorized to perform this action",
+		})
+	}
+
 	if err := c.BodyParser(&record); err != nil {
 		log.Info(err)
 		return err
@@ -102,6 +112,15 @@ func RecordRender(c *fiber.Ctx) error {
 func RecordDelete(c *fiber.Ctx) error {
 
 	id := c.Params("id")
+
+	userCredentials := c.Locals("user").(fiber.Map)
+	role := userCredentials["Role"].(string)
+
+	if role != "tutor" {
+		return c.Status(http.StatusUnauthorized).JSON(fiber.Map{
+			"error": "unauthorized to perform this action",
+		})
+	}
 	result := initializers.DB.Delete(&models.Record{}, id)
 	if result.Error != nil {
 		return c.Status(http.StatusBadRequest).JSON(
